Relocated the methods in brave_vpn/pref_names.cc
fix https://github.com/brave/brave-browser/issues/24653
