+++ components/brave_vpn/brave_vpn_utils.cc
@@ -9,8 +9,12 @@
 #include "base/notreached.h"
 #include "brave/components/brave_vpn/brave_vpn_constants.h"
 #include "brave/components/brave_vpn/features.h"
+#include "brave/components/brave_vpn/pref_names.h"
+#include "brave/components/p3a_utils/feature_usage.h"
 #include "brave/components/skus/browser/skus_utils.h"
 #include "brave/components/skus/common/features.h"
+#include "build/build_config.h"
+#include "components/prefs/pref_registry_simple.h"
 
 namespace brave_vpn {
 
@@ -29,6 +33,28 @@ std::string GetManageUrl(const std::string& env) {
 
   NOTREACHED();
   return brave_vpn::kManageUrlProd;
+}
+
+void RegisterProfilePrefs(PrefRegistrySimple* registry) {
+#if !BUILDFLAG(IS_ANDROID)
+  registry->RegisterBooleanPref(prefs::kBraveVPNShowButton, true);
+  registry->RegisterListPref(prefs::kBraveVPNRegionList);
+  registry->RegisterStringPref(prefs::kBraveVPNDeviceRegion, "");
+  registry->RegisterStringPref(prefs::kBraveVPNSelectedRegion, "");
+  registry->RegisterBooleanPref(prefs::kBraveVPNShowDNSPolicyWarningDialog,
+                                true);
+#elif BUILDFLAG(IS_ANDROID)
+  registry->RegisterStringPref(prefs::kBraveVPNPurchaseTokenAndroid, "");
+  registry->RegisterStringPref(prefs::kBraveVPNPackageAndroid, "");
+#endif
+  registry->RegisterStringPref(prefs::kBraveVPNEEnvironment,
+                               skus::GetDefaultEnvironment());
+}
+
+void RegisterLocalStatePrefs(PrefRegistrySimple* registry) {
+  p3a_utils::RegisterFeatureUsagePrefs(
+      registry, prefs::kBraveVPNFirstUseTime, prefs::kBraveVPNLastUseTime,
+      prefs::kBraveVPNUsedSecondDay, prefs::kBraveVPNDaysInMonthUsed);
 }
 
 }  // namespace brave_vpn
