+++ components/brave_vpn/pref_names.cc
@@ -5,10 +5,6 @@
 
 #include "brave/components/brave_vpn/pref_names.h"
 
-#include "brave/components/p3a_utils/feature_usage.h"
-#include "brave/components/skus/browser/skus_utils.h"
-#include "components/prefs/pref_registry_simple.h"
-
 namespace brave_vpn {
 namespace prefs {
 
@@ -17,27 +13,6 @@ const char kBraveVPNPurchaseTokenAndroid[] =
     "brave.brave_vpn.purchase_token_android";
 const char kBraveVPNPackageAndroid[] = "brave.brave_vpn.package_android";
 #endif
-
-void RegisterProfilePrefs(PrefRegistrySimple* registry) {
-#if !BUILDFLAG(IS_ANDROID)
-  registry->RegisterBooleanPref(kBraveVPNShowButton, true);
-  registry->RegisterListPref(kBraveVPNRegionList);
-  registry->RegisterStringPref(kBraveVPNDeviceRegion, "");
-  registry->RegisterStringPref(kBraveVPNSelectedRegion, "");
-  registry->RegisterBooleanPref(kBraveVPNShowDNSPolicyWarningDialog, true);
-#elif BUILDFLAG(IS_ANDROID)
-  registry->RegisterStringPref(kBraveVPNPurchaseTokenAndroid, "");
-  registry->RegisterStringPref(kBraveVPNPackageAndroid, "");
-#endif
-  registry->RegisterStringPref(kBraveVPNEEnvironment,
-                               skus::GetDefaultEnvironment());
-}
-
-void RegisterLocalStatePrefs(PrefRegistrySimple* registry) {
-  p3a_utils::RegisterFeatureUsagePrefs(
-      registry, kBraveVPNFirstUseTime, kBraveVPNLastUseTime,
-      kBraveVPNUsedSecondDay, kBraveVPNDaysInMonthUsed);
-}
 
 }  // namespace prefs
