+++ components/brave_vpn/pref_names.h
@@ -8,8 +8,6 @@
 
 #include "build/build_config.h"
 
-class PrefRegistrySimple;
-
 namespace brave_vpn {
 namespace prefs {
 
@@ -32,9 +30,6 @@ constexpr char kBraveVPNLastUseTime[] = "brave.brave_vpn.last_use_time";
 constexpr char kBraveVPNUsedSecondDay[] = "brave.brave_vpn.used_second_day";
 constexpr char kBraveVPNDaysInMonthUsed[] =
     "brave.brave_vpn.days_in_month_used";
-
-void RegisterProfilePrefs(PrefRegistrySimple* registry);
-void RegisterLocalStatePrefs(PrefRegistrySimple* registry);
 
 }  // namespace prefs
